refactor: drop package-level type stub from @vivliostyle/remark-ruby
