+++ packages/remark-ruby/package.json
@@ -23,6 +23,7 @@
   },
   "dependencies": {
     "@types/mdast": "^3.0.0",
+    "@types/unist": "^2.0.0",
     "mdast-util-to-hast": "^11.3.0",
     "unified": "^9.2.0",
     "unist-builder": "^3.0.1"
@@ -30,7 +31,7 @@
   "devDependencies": {
     "rehype-stringify": "^8.0.0",
     "remark-breaks": "^1.0.5",
-    "remark-parse": "^8.0.2",
+    "remark-parse": "^8.0.3",
     "remark-rehype": "^8.1.0"
   }
 }
