+++ packages/remark-ruby/tests/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "../tsconfig.json",
-  "include": [".", "../types"],
+  "include": ["."],
   "exclude": [],
   "compilerOptions": {
     "rootDir": "..",
