+++ packages/remark-ruby/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "../../tsconfig.json",
-  "include": ["src", "types"],
+  "include": ["src"],
   "exclude": ["tests"],
   "compilerOptions": {
     "rootDir": "./src",
